A voxel based cube display
This uses the matplotlib's inbuilt voxels(3D pixels) to create a 3D cube. However because we are using voxels, the cube is made up of 3x3x3 smaller cubes all of the same color. This means corner pieces will be the correct color. I added this to create something that resembles a cube. I don't think we should peruse this approach